Refactor: Extract `appResource.ts`
